refactor(web): move tenderly utils to @safe-global/utils
